test_requirements.txt
coverage
flake8
ipython
ipdb
pytest 
pytest-cov 
charms.reactive
